+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/MxSwiftMessage.java
@@ -237,16 +237,24 @@ public class MxSwiftMessage extends AbstractSwiftMessage {
     }
 
     private void _updateFromMessage(final MxId id, final MessageMetadataStrategy metadataStrategy) {
-        if (message() != null && message().length() > 0) {
+        if (message() != null && !message().isEmpty()) {
             MxId identifier = id != null
                     ? id
                     : MxParseUtils.identifyMessage(this.message()).orElse(null);
-            extractMetadata(identifier, getAppHdr(), metadataStrategy);
+
+            // the identifyMessage above will also attempt the header but with a generic parsing, so we try again
+            // with a specific header parsing if the identifier is still null
+            AppHdr header = getAppHdr();
+            if (identifier == null && header != null && header.messageName() != null) {
+                identifier = new MxId(header.messageName());
+            }
+
+            extractMetadata(identifier, header, metadataStrategy);
         }
     }
 
     /**
-     * Updates the the attributes with the raw message and its metadata from the given raw (XML) message content.
+     * Updates the attributes with the raw message and its metadata from the given raw (XML) message content.
      *
      * @param mx the new message content
      * @see #updateFromMessage()