+++ iso20022-core/src/test/java/com/prowidesoftware/swift/model/mx/MxParseUtilsTest.java
@@ -86,4 +86,76 @@ public class MxParseUtilsTest {
         Optional<MxId> id = MxParseUtils.identifyMessage(xml);
         assertTrue(id.isPresent());
     }
+
+    @Test
+    public void testIdentifyMessage_FromBAH() {
+        final String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>"
+                + "<Message>\n"
+                + "<AppHdr xmlns=\"urn:iso:std:iso:20022:tech:xsd:head.001.001.01\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\">\n"
+                + "<Fr> \n"
+                + "	<FIId>\n"
+                + "		<FinInstnId>\n"
+                + "			<BICFI>FOOCUS3NXXX</BICFI>\n"
+                + "			<ClrSysMmbId>\n"
+                + "				<ClrSysId>\n"
+                + "					<Prtry>T2S</Prtry>\n"
+                + "				</ClrSysId>\n"
+                + "				<MmbId>ADMNUSERLUXCSDT1</MmbId>\n"
+                + "			</ClrSysMmbId>\n"
+                + "			<Othr>\n"
+                + "				<Id>FOOTXE2SXXX</Id>\n"
+                + "				</Othr> \n"
+                + "		</FinInstnId> \n"
+                + "	</FIId> \n"
+                + "</Fr> \n"
+                + "<To> \n"
+                + "	<FIId>\n"
+                + "		<FinInstnId>\n"
+                + "			<BICFI>ABICUS33</BICFI>\n"
+                + "			<Othr>\n"
+                + "				<Id>AARBDE5W100</Id>\n"
+                + "			</Othr>\n"
+                + "		</FinInstnId> \n"
+                + "	</FIId> \n"
+                + "</To> \n"
+                + "<BizMsgIdr>2012111915360885</BizMsgIdr>\n"
+                + "<MsgDefIdr>seev.031.002.03</MsgDefIdr> \n"
+                + "<BizSvc>CSD</BizSvc> \n"
+                + "<CreDt>2015-08-27T08:59:00Z</CreDt>\n"
+                + "</AppHdr>\n"
+                + "<Document></Document>\n"
+                + "</Message>";
+        MxId id = MxParseUtils.identifyMessage(xml).orElse(null);
+        assertNotNull(id);
+        assertEquals("seev.031.002.03", id.id());
+    }
+
+    @Test
+    public void testIdentifyMessage_LegacyBAH() {
+        final String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>"
+                + "<Message>\n"
+                + "<h:AppHdr xmlns:h=\"urn:iso:std:iso:20022:tech:xsd:head.001.001.01\">" + "   <h:Fr>"
+                + "      <h:FIId>"
+                + "         <h:FinInstnId>"
+                + "            <h:Nm>Not available</h:Nm>"
+                + "         </h:FinInstnId>"
+                + "      </h:FIId>"
+                + "   </h:Fr>"
+                + "   <h:To>"
+                + "      <h:FIId>"
+                + "         <h:FinInstnId>"
+                + "            <h:Nm>Not available</h:Nm>"
+                + "         </h:FinInstnId>"
+                + "      </h:FIId>"
+                + "   </h:To>"
+                + "   <h:BizMsgIdr>AAAAAAAAAA222222</h:BizMsgIdr>"
+                + "   <h:MsgDefIdr>seev.037.002.02</h:MsgDefIdr>"
+                + "   <h:CreDt>2017-08-08T16:58:01Z</h:CreDt>"
+                + "</h:AppHdr>"
+                + "<Document></Document>\n"
+                + "</Message>";
+        MxId id = MxParseUtils.identifyMessage(xml).orElse(null);
+        assertNotNull(id);
+        assertEquals("seev.037.002.02", id.id());
+    }
 }